remove setuputils, merge functionality into libutils
